Fixs for notifications
FIX https://github.com/gitskarios/Gitskarios/issues/202
FIX https://github.com/gitskarios/Gitskarios/issues/195

+++ src/main/java/com/alorma/github/sdk/bean/dto/response/Notification.java
@@ -18,7 +18,7 @@ public class Notification extends ShaUrl implements Comparable<Notification>{
 	public String updated_at;
 	public String last_read_at;
 	
-	public int adapter_repo_parent_id;
+	public Long adapter_repo_parent_id;
 
 	@Override
 	public int compareTo(Notification another) {
